fix refactor error
