+++ app/src/main/java/com/nermeen/movie_app/ui/home/adapter/MovieAdapter.kt
@@ -1,9 +1,8 @@
-package com.nermeen.movie_app.ui.home
+package com.nermeen.movie_app.ui.home.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.nermeen.movie_app.ui.home.viewModel.HomeViewModel
-import com.nermeen.movie_app.ui.home.adapter.MoviesDiffCallback
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
